KAS-2308 Refactoring placement of code and files
